mavlink: move BATTERY_STATUS to separate stream header
